Refactor: rename legacy parser to LegacyParser
